requirements.txt
Django==1.11
djangorestframework==3.8.2

